+++ CHANGES.rst
@@ -1,6 +1,23 @@
 CHANGE LOG
 ==========
 
+5.4.2
+-----
+- Bug Fix: show an error on frontend if the user did answer the security question on registration when they try to reset their password
+- Won't be able to delete a security question when users have used it
+
+5.4.1
+-----
+- Bug Fix: enable search via username on permissions
+
+5.4.0
+-----
+- Add site permissions to admin users
+
+5.3.3
+-----
+- Bug Fix: Filter Security Questions by site
+
 5.3.2
 -----
 - Bug Fix: Add null check in backend for CAS